+++ src/crypto/keys.rs
@@ -20,8 +20,8 @@ use super::signature::Signature;
 /// of 2048 bits. However, as that might change in the future, we are not
 /// hard-coding that format but rather use this type – which for the time
 /// being is zero-sized.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct PublicKeyFormat;
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
+pub struct PublicKeyFormat(());
 
 /// # ASN.1 Algorithm Identifiers
 ///
@@ -59,7 +59,7 @@ impl PublicKeyFormat{
     ) -> Result<Self, S::Err> {
         oid::RSA_ENCRYPTION.skip_if(cons)?;
         cons.take_opt_null()?;
-        Ok(PublicKeyFormat)
+        Ok(PublicKeyFormat::default())
     }
 
     /// Provides an encoder for the algorihm identifier.
@@ -160,12 +160,6 @@ impl PublicKey {
 
     pub fn to_subject_name(&self) -> Name {
         Name::from_captured(self.encode_subject_name().to_captured(Mode::Der))
-    }
-}
-
-impl AsRef<[u8]> for PublicKey {
-    fn as_ref(&self) -> &[u8] {
-        self.bits()
     }
 }