+++ Changelog.md
@@ -8,6 +8,14 @@ Breaking
 * Restructured how certificate’s SubjectAccessInfo is parsed and accessed.
   It now keeps the first mandatory URI of the four relevant access methods
   readily available. [(#34)]
+* `Cert` has completely changed, `CertBuilder` is gone and has been
+  replace by `TbsCert` which can be used for building. [(#XX)]
+* `cert::Validity`, `x509::Name`, `x509::SignedData`: `encode` renamed to
+  `encode_ref` to comply with standard naming scheme. [(#XX)]
+* `DigestAlgorithm`, `PublicKeyFormat`, and `SignatureAlgorithm` are not
+  unit structs anymore. They impl `Default` which should be used to get
+  the recommended (read: only available) values. This is so we can
+  transparently change them into enums later on if necessary. [(#XX)]
 
 New